重构 IO
