+++ libtpms/TPMS/CustomTpmsSingleSurfaceAlgorithm.cpp
@@ -1,0 +1,6 @@
+#include "CustomTpmsSingleSurfaceAlgorithm.h"
+
+CustomTpmsSingleSurfaceAlgorithm::CustomTpmsSingleSurfaceAlgorithm()
+{
+
+}
