+++ libtpms/TPMS/CustomTpmsSingleSurfaceAlgorithm.h
@@ -1,0 +1,11 @@
+#ifndef CUSTOMTPMSSINGLESURFACEALGORITHM_H
+#define CUSTOMTPMSSINGLESURFACEALGORITHM_H
+
+
+class CustomTpmsSingleSurfaceAlgorithm
+{
+public:
+    CustomTpmsSingleSurfaceAlgorithm();
+};
+
+#endif // CUSTOMTPMSSINGLESURFACEALGORITHM_H
