+++ libtpms/TPMS/CustomTpmsSingleSurfaceConfig.cpp
@@ -1,0 +1,1 @@
+#include "CustomTpmsSingleSurfaceConfig.h"
